Create callback.php
This is your callback script. The URL for this file should be your redirect URI. This script should receive the authorization code and be securely stored so you can exchange it for an access token. The process of exchanging the authorization code for an access token is already part of the file 1_initail_authorzation.py